retoques

+++ backend/RealEstateManagement/RealEstateManagement/Repository/ProductoRepository.cs
@@ -61,6 +61,10 @@ namespace RealEstateManagement.Repository
         public void RemoveProducto(Guid codigoProducto)
         {
             var producto = context.Productos.FirstOrDefault(p => p.Codigo == codigoProducto) ?? throw new Exception($"El producto con codigo {codigoProducto} no existe");
+            if(producto.EstadoProducto ! == 0)
+            {
+                throw new Exception($"El producto no puede ser borrado,solo los productos con estado disponible pueden borrarse.");
+            }
             context.Productos.Remove(producto);
 
             context.SaveChanges();
